// apps/ui/src/lib/api-client/models/AccessHealthResponseCapabilities.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.5-alpha.202604142009.f75bd474
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * Provider capabilities
 * @export
 * @interface AccessHealthResponseCapabilities
 */
export interface AccessHealthResponseCapabilities {
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    groups: boolean;
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    members: boolean;
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    sync: boolean;
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    events: boolean;
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    groupDoors: boolean;
    /**
     * @type {boolean}
     * @memberof AccessHealthResponseCapabilities
     */
    realtime: boolean;
}

/**
 * Check if a given object implements the AccessHealthResponseCapabilities interface.
 */
export function instanceOfAccessHealthResponseCapabilities(value: object): value is AccessHealthResponseCapabilities {
    if (!('groups' in value) || value['groups'] === undefined) return false;
    if (!('sync' in value) || value['sync'] === undefined) return false;
    if (!('groupDoors' in value) || value['groupDoors'] === undefined) return false;
    if (!('events' in value) || value['events'] === undefined) return false;
    if (!('realtime' in value) || value['realtime'] === undefined) return false;
    if (!('members' in value) || value['members'] === undefined) return false;
    return true;
}

export function AccessHealthResponseCapabilitiesFromJSON(json: any): AccessHealthResponseCapabilities {
    return AccessHealthResponseCapabilitiesFromJSONTyped(json, false);
}

export function AccessHealthResponseCapabilitiesFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccessHealthResponseCapabilities {
    if (json == null) {
        return json;
    }
    return {
        
        'groups': json['groups'],
        'members': json['members'],
        'sync': json['sync'],
        'events': json['events'],
        'groupDoors': json['groupDoors'],
        'realtime': json['realtime'],
    };
}

export function AccessHealthResponseCapabilitiesToJSON(json: any): AccessHealthResponseCapabilities {
    return AccessHealthResponseCapabilitiesToJSONTyped(json, false);
}

export function AccessHealthResponseCapabilitiesToJSONTyped(value?: AccessHealthResponseCapabilities | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'groups': value['groups'],
        'members': value['members'],
        'sync': value['sync'],
        'events': value['events'],
        'groupDoors': value['groupDoors'],
        'realtime': value['realtime'],
    };
}