// apps/ui/src/lib/api-client/models/SyncResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.5-alpha.202604151658.c3071eb4
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SyncResponseFailedInner } from './SyncResponseFailedInner';
import {
    SyncResponseFailedInnerFromJSON,
    SyncResponseFailedInnerFromJSONTyped,
    SyncResponseFailedInnerToJSON,
    SyncResponseFailedInnerToJSONTyped,
} from './SyncResponseFailedInner';

/**
 * @export
 * @interface SyncResponse
 */
export interface SyncResponse {
    /**
     * Emails of newly created members
     * @type {Array<string>}
     * @memberof SyncResponse
     */
    created: Array<string>;
    /**
     * Emails already existing
     * @type {Array<string>}
     * @memberof SyncResponse
     */
    skipped: Array<string>;
    /**
     * Emails that failed to sync
     * @type {Array<SyncResponseFailedInner>}
     * @memberof SyncResponse
     */
    failed: Array<SyncResponseFailedInner>;
}

/**
 * Check if a given object implements the SyncResponse interface.
 */
export function instanceOfSyncResponse(value: object): value is SyncResponse {
    if (!('failed' in value) || value['failed'] === undefined) return false;
    if (!('created' in value) || value['created'] === undefined) return false;
    if (!('skipped' in value) || value['skipped'] === undefined) return false;
    return true;
}

export function SyncResponseFromJSON(json: any): SyncResponse {
    return SyncResponseFromJSONTyped(json, false);
}

export function SyncResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): SyncResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'created': json['created'],
        'skipped': json['skipped'],
        'failed': ((json['failed'] as Array<any>).map(SyncResponseFailedInnerFromJSON)),
    };
}

export function SyncResponseToJSON(json: any): SyncResponse {
    return SyncResponseToJSONTyped(json, false);
}

export function SyncResponseToJSONTyped(value?: SyncResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created': value['created'],
        'skipped': value['skipped'],
        'failed': value['failed'] == null ? undefined : ((value['failed'] as Array<any>).map(SyncResponseFailedInnerToJSON)),
    };
}